Refactor WebView logic out of MainActivity
